requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.6.1
alembic==1.14.0
python-dotenv==1.0.1
